refactor: split AgentExecutionContextFactory into partials by responsibility
The factory was 666 lines, past this repo's 400-line cap (issue #251). Split
along the four jobs it actually does, following the PlanExecutor.* convention:

  .Prompt.cs           authoritative static system prompt composition
  .SkillDisclosure.cs  progressive-disclosure budget charging + fallback report
  .ContextProviders.cs the ordered AIContextProvider rail
  .Resolution.cs       deployment/framework/tool-ceiling/middleware resolvers

Pure move: the member signature set is byte-identical before and after, and no
body was edited. Largest file is now 310 lines. The two positional invariants on
the provider rail (tool-contributing providers above the permission filter, the
per-turn measurer last) are restated on the partial that owns them, since that
file is now where someone edits the order.
